Remove `label`
IMO, Label is an unnecessary complication.

* Label has no actual use on the client.

* Because labels cannot collide, the key generator needs to either keep some state across all invocations (you need a trustable store), or generate a UUID (it's just a random junk) to be used as a label.

* A TLS server is required to at first compare the label and then check the digest.

These three issues go away if we eliminate the label. Key generators do not need to care about uniqueness. TLS servers can just look for the matching digest and that's it.

Note that a server operator can still associate an label to the ESNIKeys structure for internal use. The private key will be kept in relation to the structure anyways. So having a private label does not complicate things at all.