Improve ESNI robustness.
Closes #82.

ESNI has different compatibility characteristics than most TLS
extensions. It involves client state, and encrypting the SNI makes the
ClientHello unusable for servers that do not have the key or do not
implement ESNI. This breaks common rollout and rollback strategies and
places strong requirements on DNS/server consistency. This makes
deploying ESNI risky for servers.

This change aims to reduce these deployment risks. ESNI is much more
useful if widely deployed. Moreover, if ESNI is unreliable due to
complex requirements on the server, clients will be disincentivized to
implement it. The DNS/server consistency requirements also
disincentivize rotating keys (section 7.2.5).

This change introduces the notion of a "public name" which allows the
client to receive authenticated signals from the server that the
ESNIKeys are out of sync, so clients can safely repair key mismatches.
It also aligns with existing server behavior before ESNI, so partial
rollouts and rollbacks can also be handled.

The fix for the latter also handles MITM scenarios, replacing the
existing client recommendation:

   A Web client client can securely detect case (2) because it will
   result in a connection which has an invalid identity (most likely)
   but which is signed by a certificate which does not chain to a
   publicly known trust anchor.  The client can detect this case and
   disable ESNI while in that network configuration.

   https://tools.ietf.org/html/draft-ietf-tls-esni-02#section-6.2

The public name notion fixes a bug in that text: an attacker could
disable ESNI on all servers, given the existence of *any* such server.