Clarify how anti-replay protection is achieved
The SNI encryption key is not bound to the Client Hello, one could for
example "cut and paste" the ESNI value in a different Client Hello with
the same Random field, key_share extension and ESNI extension, but with
a different, say, supported_groups extension.

Remove the (EC)DH argument since it only ensures that an attacker cannot
compute the SNI encryption key, but actual anti-replay protection is
provided by binding to the key_share extension.