Advertise multiple ESNIConfigs as a unit.
(This was split out of
https://github.com/tlswg/draft-ietf-tls-esni/pull/199 as the hopefully
less controversial idea.)

We have two levels of mechanisms for servers to advertise parameters.
Each ESNIConfig has multiple keys, so a server can suport multiple
groups at once. These keys are stapled together so a single HTTPSSVC
record delivers the entire configuration.

Then we have a versions/extensions mechanism. A server supporting
multiple versions of ESNI or multiple sets of mandatory extensions must
advertise multiple ESNIConfigs. Unlike the group advertisement, these
require separate HTTPSSVC records. This means the full ESNI config is
not reliably available to logic across different layers of the client
stack. In particular, the HTTPSSVC spec text in [0] does not work.

Currently, a single ESNIConfig, rather than a set of them, is correlated
with an IP address or CNAME, by way of the HTTPSSVC record. A server
which supports multiple ESNI versions or mandatory extension sets must
advertise several HTTPSSVC records, potentially with different IPs. DNS
and transport logic must know TLS capabitilies to pick the right
HTTPSSVC record.

This PR introduces an ESNIConfigs structure which advertises a list of
ESNIConfig structures. The ESNIConfigs is the new unit of ESNI
configuration. This should be paired with a change to
draft-nygren-dnsop-svcb-httpssvc to carry an ESNIConfigs, not an
ESNIConfig (formerly ESNIKeys).

Now the server can include the full config at once, so the DNS and
transport logic can reasonably pick an HTTPSSVC purely based on whether
the TLS stack supports ESNI, rather than the full set of capabilities.
This fixes the text in [0].

That ESNIConfigs and ESNIConfig are separate structures is slightly
goofy naming. Alternate names if folks prefer:

- ESNIConfig/ESNIConfigs => ESNIConfig/ESNIBundle

- ESNIConfig/ESNIConfigs => ESNIKeys/ESNIConfig
  (Nuisance: an ESNIConfig is a list of ESNIKeyses, which brings back
  the double plural.)

(While I'm here, fix the remnants of ESNIKeyShares. I'll split that out
if this PR is controversial, but GitHub is really bad at dependent PRs,
so I've left it together for now.)

[0] https://tools.ietf.org/html/draft-nygren-dnsop-svcb-httpssvc-00#section-8.1.1