Unify version/extension and key share negotiation mechanisms.
We currently have two levels of mechanisms for servers to advertise
multiple sets of parameters. First, each ESNIConfig has multiple keys in
it, so a server can suport multiple groups at once. These keys are
stapled together so a single HTTPSSVC record, etc., delivers the entire
configuration.

At a layer above that, we have a version/extension-based mechanism. The
server may advertise multiple ESNIConfigs to cover multiple versions of
ESNI (hopefully we won't need this mechanism after standardization, but
it's quite useful for draft versions) or multiple sets of mandatory
extensions. Unlike the group advertisement, these are each in separate
configuration units, so, e.g., they may be in multiple HTTPSSVC records.

This PR flattens the two mechanisms into one. Each ESNIConfig now has a
single KeyShareEntry. Multiple groups and multiple versions are
expressed in the same way: multiple ESNIConfig structures. It then
introduces a wrapper ESNIConfigs structure which is the new unit of ESNI
configuration delivered in HTTPSSVC, etc.

This should be paired with a change to draft-nygren-dnsop-svcb-httpssvc
to carry an ESNIConfigs, not an ESNIConfig (formerly ESNIKeys).

This fixes a number of oddities:

- Client and server need to maintain two levels of configuration. On the
  server, the structure has two layers of lists, something like
  list(tuple(ESNIConfig, list(PrivateKey))). Now the server maintains
  list(tuple(ESNIConfig, PrivateKey)). Conversely, the client needed two
  negotiation loops whereas now it only requires one.

- Logic across different layers of the client stack now more reliably
  have the full ESNI configuration available in one unit.

  In the old design, a single ESNIConfig, rather than a set of them,
  would be correlated with an IP address or CNAME, by way of the
  HTTPSSVC record. A server which supportsed multiple ESNI versions or
  mandatory extension sets would advertise several HTTPSSVC records,
  potentially with different IPs. That means the DNS and transport
  connection logic must know about TLS capabitilies to pick the right
  HTTPSSVC record.

  In the new design, the server includes all its variations at once, so
  the DNS and transport logic can reasonably pick an HTTPSSVC purely
  based on whether the TLS stack supports ESNI, rather than the full set
  of capabilities. That is, the text in [0] does not work right now, and
  this PR fixes it.

The downside is that, in the likely case that the server only supports
one version and one set of extensions, the non-key ESNI config bits
(notably the public name; the rest is small) are repeated for every
named group. However, the public name should not be that large and
servers should only need one or two groups.

This also does make ESNIConfigs and ESNIConfig separate structures,
which is slightly goofy. Alternate names if folks prefer:

- ESNIConfig/ESNIConfigs => ESNIConfig/ESNIBundle
- ESNIConfig/ESNIConfigs => ESNIKey/ESNIConfig

[0] https://tools.ietf.org/html/draft-nygren-dnsop-svcb-httpssvc-00#section-8.1.1
[1] https://www.imperialviolet.org/2016/05/16/agility.html