Add a length prefix to ESNIConfig contents.
ESNIConfig has a version number in front for extensibility, but it is
unclear what the version invariants are. Presumably we would like
versions to change the contents of an ESNIConfig (indeed different
drafts have done so), but it is not obvious this is possible and it is
impossible to skip over an ESNIConfig with a version you do not
understand.

Fix this by moving the body of an ESNIConfig into an ESNIConfigContents
and adding a length prefix in front.

Fixes #192.