Flatten ESNI negotiation mechanisms.
(This was split out of
https://github.com/tlswg/draft-ietf-tls-esni/pull/199 as the possibly
more controversial idea. It's a simplification but does make configs
larger.)

We have two layers of parameter negotiation. Multiple KeyShareEntry
values in an ESNIConfig, and multiple ESNIConfig values in an
ESNIConfigs. We only need one mechanism, so store only one KeyShareEntry
in an ESNIConfig.

This simplifies representing client and server configuration, which
simplies implementation internals as well as interchange formats. Right
now, on the server, configuration has two layers of lists, something
like list(tuple(ESNIConfig, list(PrivateKey))). Now the server maintains
list(tuple(ESNIConfig, PrivateKey)).

Note also that draft-farrell-tls-pemesni-00 has this oddity around what
to do if the ESNIKeys contains multiple keys. This is particularly a
nuisance because a single standalone PRIVATE KEY / ESNIKEY pair is not
usable because the server has committed to answering to *all* keys in
the ESNIKeys. The server must consume multiple PEM files, match up the
ESNIKeyses, and then assert that it has a complete set.

The downside is that, in the likely case that the server only supports
one version and one set of extensions, the non-key ESNI config bits
(notably the public name; the rest is small) are repeated for every
named group. However, the public name should not be that large and
servers should only need one or two groups (see
https://www.imperialviolet.org/2016/05/16/agility.html).