Fix superfluous padding edge cases.
Consider the following values for L and D, where L is the server's hint
and D is the length of the domain name.

    D1     D2      D3
... -+--|---+---|---+---|----->
        L     32*n   32*(n+1)

This change pads D according to the following rules.

- D1: D <= L, so round up to L.
- D2: 32*(n-1) <= L < D <= 32*n, round up to 32*(n+1).
- D3: 32*(n-1) <= L <= 32*n < D, round up to 32*(n+1).

Case D2 exists because the anonymity set of of D, if rounded to 32*n
(instead of 32*(n+1), is less than 32. The new padding algorithm ensures
that the anonymity set for D whenever D > L is at least 32 elements.