Fix ECH and version negotiation order.
The version check on the outer ClientHello doesn't make much sense
since, when ECH is accepted, the version is based on the inner
ClientHello anyway. See
https://github.com/tlswg/draft-ietf-tls-esni/issues/309#issuecomment-702233855.

I've left a TODO since it seems weird to have a client requirement like
that be so load-bearing without a corresponding server check. Not sure
where we should go there.