Clarify client handling of retry keys with GREASE.
The intent was that the server always sends the retry keys (it cannot
distinguish stale keys from GREASE when the client is actually trying to
connect to the public name), while the client just ignores them, but I
did a poor job of describing this, so add a key sentence.

This addresses #295.