Unify trial decryption and config_id modes.
The trial decryption mode is specified oddly. It says to use trial
decryption on the server, but the process following it only works in the
config_id mode (it says decryption failure is fatal). It also behaves
differently because decryption failure is an ECH rejection rather than a
fatal error.

Unify them by instead saying the server gathers a (possibly singleton)
set of candidate ECHConfig values and then tries them in succession.
Note this aligns the config_id mode's error handling with the trial
decryption mode, since the other direction is not possible. (So it goes
the other direction on #290.)

This should tidy up the oddity in
https://github.com/tlswg/draft-ietf-tls-esni/pull/339#discussion_r505718046,
and meshes well with GREASE and ClientHelloOuterAAD.

The client text also contradicts this, so add a few more sentences.