Clarify client and server roles
Whether the client offers ECH or sends GREASE ECH depends on whether it
can find a suitable ECH config. Whether the server acts like the
client-facing server or the backend server depends on whether it has
custody of a set of ECH configs. This change clarifies these decision
making processes.

As a bonus, this change also cleans up the section titles and levels in
a bit.

Co-authored-by: Christopher Wood <caw@heapingbits.net>