Split-mode HRR is just hard
...but don't give both servers a cookie.

That is, if a client-facing server forwards the ClientHello, it can't do
so statelessly without coordinating with the backend.  This is a
consequence of moving to a single HelloRetryRequest, but it's a minor
one.  In most cases, there will be state established anyway just to get
the connection forwarded.  DTLS might not work that way, but it's still
likely reasonable to assume state.

Closes #333.