Relax computation of the acceptance signal.
We previously derived the acceptance signal from the handshake secret.
This meant that clients which used the wrong ECHConfig might need to
process ServerHello extensions twice before computing the sigal, which
can be problematic for some libraries. Given that the signal's secrecy
is entirely dependent on ClientHelloInner.random, we can relax the signal
computation and base it on the transcript alone, which includes
ClientHelloInner.random, rather than a secret derived from that transcript.

Closes #399.