Confirm ECH acceptance after HRR
If an HRR is sent in response to CH, the client doesn't know whether ECH
was accepted. As a result, it doesn't know whether it's CHI or CHO that
needs to be updated. The spec attempts to address this ambiguity by
requiring that all HRR-sensitive parameters are the same in CHI and CHO.
This criterion is currently underspecified, and it's not clear how to
make sure that future extensions don't violate it. (See issue #374.)

This change solves this ambiguity by adding an explicit signal of ECH
acceptance to HRR. This violates "don't stick out" whenever an HRR is
inevitable for a particular client and server configuration. However,
this change is designed to prevent an active attacker from
distinguishing real ECH from GREASE ECH by triggering HRR.