Relax public_name to public_identity and describe how clients construct CHOuter based on it.
Previously, public_name was assumed to be a DNS name. #413 added validation criteria
for this value. This change relaxes public_name to be a reference identity for the
client-facing server. As a result, clients use this identity when constructing
CHOuter and authenticating the client-facing server. This identity may be either
a DNS name or an IP address, depending on the particular client-facing server
deployment.