Rephrase maximum_name_length description.
While padding is "described below", it's pretty far away now. We also
recommend the padding scheme whether or not the value is non-zero. The
text about wildcard names also applies in both cases and seems to
describe when the server name may exceed the max.

This rephases things a bit.