Revise computation of acceptance signal
We currently use the all-zero string as the PRK input of
HKDF-Expand-Label(). This usage of this function isn't generically safe,
as PRK is required to be pseudorandom (see RFC5869). In particular,
although the transcript contains a pseudorandom string
(ClientHelloInner.random), passing the hash of the transcript as the
info string may not be sufficient to guarantee that the output OKM is
pseudorandom (or even unpredictable).

This changes the acceptance signal to use HKDF-Extract(0,
ClientHelloInner.random) as the PRK.