Describe the whole procedure more precisely
First, this describes how to compute payload length. It is sadly a
little vague because RFC5116 doesn't give a name to this algorithm,
merely that:

   Each AEAD algorithm MUST provide a description relating the length of
   the plaintext to that of the ciphertext.

Second, this fixes both the first and second ClientHelloOuter
construction algorithms. In both the old and new construction, it was
necessary for the client to compute ClientHelloOuterAAD before
ClientHelloOuter, but the process describes it in reverse. Describe it
in the correct order by pulling the encryption procedure into a

Third, this makes the second ClientHelloOuter logic a bit more precise
by making the encryption procedure work for both.

Finally, this fixes some minor issues: a sentence about PSKs was not
updated for GREASE PSKs, and I've moved "Unless specified by the
application using (D)TLS or externally configured on both sides" to the
optional configs section, so it can apply to both client and server.