Expand on config_ids.
The discussion on how to allocate config_ids made it very unclear
whether we cared about collisions in DNS or across client-facing
servers. (The former is insufficient and the latter is practically
impossible.) Instead, pull this into a separate section, so we can be
clear this is a recommendation for the client-facing server and the
backend server.

Also downgrade from MUST to SHOULD because collisions work fine, just
don't perform as well.

(It was also weird that we talked about config_id in not the section
about config_ids, but in the little side note about HpkeKeyConfig.)