Move name validation to the authentication section
This is partly motivated by my interests in doing something evil,
but mostly this is because coupling name validity to ECH config
validity is a layering violation.  It's fine to invoke some name
validation, but that should be dictated by the needs of the thing that
ends up relying on that name.

This corrects both problems.

In doing this, I realized that RFC 791 says nothing about the IP address
textual format.  That's problematic, but I couldn't come up with
anything better in short notice.

Closes #628.
Closes #637.