This addresses the issues in Stewart Bryant's review. (#645)
The final issue asked if we should replace the reference to RFC 5077
to RFC 8446, but this text is correct because the reference is to part
of the internal example structure in 5077 and 8446 is just agnostic on
token structure. 5077 is being used by way of analogy, not as a part
of the protocol.